Refactored and cleaned up the GreedyMesher code. Need to fix chunk borders.
An easy way to fix the chunk borders is to automatically count voxels outside of the bounds of the chunk as air in CreateFaceAtPosition. This would cause some unnecessary draws.

Perhaps it would be better to delay the displaying the mesh and create an UpdateChunkBorder method. This will be required once I finish implementing removal/placement of voxels anyways.